Update swag.py
The code should run and it saves predictions (of the regular) model in
a file now.
The SWAG part *should* also work, except that I have not figured out
yet how to evaluate the swag_model Object on actual data.

`swag_model(batch["X"].to(swag_model.device))` like for EQTransformer
does not seem to work.